Creates and uses TablePanel component
